fusebox: implement ReadDir2 for the client
The existing ReadDir D-Bus method (part of the FuseBoxService D-Bus
interface) provides directory entries by calling the ReplyToReadDir
method on a *different* interface (the FuseBoxReverseService D-Bus
interface). The FBS and FBRS servers live in different processes: the
client process for one is the server process for the other.

The new ReadDir2 D-Bus method is simpler. It involves only one D-Bus
interface (FBS).

We would like to experiment with simplifying or even removing the FBRS
D-Bus interface, which would allow the FBS D-Bus implementation (i.e.
Chrome) to serve multiple clients simultaneously, not just one.

This is one half of a multi-repo change. The other half is at
https://crrev.com/c/3901035

BUG=chromium:1249754
TEST=manual

Change-Id: I317d1f32df96f1f61f878ac2834cd0bc3d04d13f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3899676
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>