camera: implement DLC download on startup
If USE_CAMERA_FEATURE_EFFECTS is set, the camera hal adapter will
download the ml-core-internal DLC and set it in the RuntimeOptions that
gets passed to the various stream manipulators.

The effects stream manipulator will check for the existing of the DLC
path before instantiating.

BUG=b:248936267
TEST=run cros_camera_service on DUT

Cq-Depend: chrome-internal:5008808
Change-Id: Id46c7568539c800ddb333f76b8938c38d65c9f6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3924636
Tested-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>