cryptohome: Connect authblock code to FingerprintManager
This change is a first attempt at pushing some of the service/state
needed for "prepare" based auth mechanisms down below the UserDataAuth
layer. It creates a "service" class that's intended to wrap some basic
interaction with the underlying FingerprintManager object that currently
does legacy fingerprint operations.

The approach here is a bit unusual in that it's somewhat both over and
under designed. It's over designed in the sense that we probably don't
need to push all this state down to the AuthBlockUtility just for legacy
fingerprint support, which isn't going to become a "real" auth block
anyway. But it's under designed in the sense that it's not defining a
full generic interface for hooking up "auth block services" either,
mostly because I don't think we have enough data yet to do this fully
properly.

I think this approach is reasonable because it's basically a preliminary
version of adding this kind of plumbing to AuthBlockUtility which moves
us in the right direction, and which can be refactored into a more
generic design, while still getting something basic working up front.

BUG=b:246824625
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Id681246358e6ad8d037ec78562ca01aaa6397852
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3905756
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>