cryptohome: Add a Verify operation to the auth factor util API.
Yes, this duplicates some of what CredentialVerifier does, in a somewhat
different way (it supports async, for one). I have already filed a bug
to do the followup work to integrate the auth factor utils and the
verifiers, but this is the easiest path to doing an async fingerprint
verify right now.

BUG=b:246824625
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I03122b6ae64e8a40e2c51961ade85d3d2d20a0d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3910716
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>