hiberman: Add EmergencyReboot cookie value
Add handling for a new possible cookie state, EmergencyReboot, that
indicates the system encountered an unrecoverable error in the abort
path at the last boot and had to reboot the system. Since we're now in
the error path of an error path, we no longer attempt to wire up
dm-snapshots for merging, but instead mostly proceed as if this was a
non-resume boot in the hope that we can recover back to a steady state.
This state can however be detected and logged.

With the existence of this state, we could end up in a situation where
the snapshot files exist, but hiberman resume-init never actually wired
them up. In this case we don't want the PendingStatefulMerge's monitor
threads to report errors, so only fire up monitor threads for snapshot
files that actually have corresponding dm targets.

Currently nobody sets this state, but we'll try to set it in the
upcoming emergency reboot path.

BUG=b:241432906
TEST=hiberman cookie -V ereboot, reboot and watch logs.

Change-Id: Ie17edf23d4b7235e4f440c95bd98731e00f1ea95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3911215
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Alexis Savery <asavery@chromium.org>