hiberman: Wire in emergency reboot
In certain failure cases, like failure to merge dm-snapshots,
failure to initiate an abort, and the complete filling of a
dm-snapshot, the safest thing to do is to reboot so that we're no longer
running on top of the failed dm-snapshots. Add in an emergency reboot
path that sets the hibernate cookie (so it can be logged upon subsequent
reboot), syncs its current logs, and reboots the system. Wire this
function up in the scenarios mentioned above.

BUG=b:241432906
TEST=Introduce bug, see emergency reboot path trigger.

Change-Id: Iacd958054e6d61171241e9f6b7e89d31270da4b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3911216
Reviewed-by: Alexis Savery <asavery@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>