shill: always compile support for Cellular
This patch removes the conditional compilation of cellular support in
the shill binary based on the "cellular" use flag.

The main benefit is the simplification of the TetheringManager code
interacting directly with shill's Cellular stack for obtaining and
monitoring an upstream network connection on Cellular.

This patch does not remove usage of the cellular flag itself in the
shill BUILD.gn file. For instance the set-apn-helper and
cellular_mobile_operator_info_tester binaries are still only compiled
and added as a dependency if "cellular" is used.

Binary size change for /usr/bin/shill:

Prior to this patch:
         |  USE="-cellular"  |  USE="cellular"
---------+-------------------+------------------
 trogdor |   1415480         |   1809480
---------+-------------------|------------------
 atlas   |   2055064         |   2633888
---------+-------------------+------------------

With this patch:
         |  USE="-cellular"  |  USE="cellular"
---------+-------------------+------------------
 trogdor |   1809480         |   1809480
---------+-------------------|------------------
 atlas   |   2633856         |   2633856
---------+-------------------+------------------

On boards were "cellular" is disabled (chromiumos-overlay embedded
profile) the additional delta is small compared to the overall footprint
of the whole cellular stack:
  MomemManager: ~1MB
  libmbim: ~0.5MB
  libqmi: ~2MB
  net-dialup/ppp: ~300kB

Most Chromebooks already include "cellular" in their builds.

BUG=b:248028325
TEST=ran shill unit tests both with and without "cellular"
 $  USE="-cellular" FEATURES=test emerge-$BOARD shill
 $  USE="cellular" FEATURES=test emerge-$BOARD shill

Cq-Depend: chromium:3949091
Change-Id: I12d610fd9509cd6a9ef84aa59826430ffdb13af4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3909969
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>