shill: Add support for VPN IPv6 in routing setup
This patch adds a new interface in the Network class to set the IPv6
link protocol properties, which can be used by the VPN service. When
both v4 and v6 properties are set, v6 properties will be used at first
to set up the Connection object, which keeps the same behavior with
other technologies.

BUG=b:251686186
BUG=b:232177767
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=tast run $DUT network.VPNConnect.*

Change-Id: Ib72b9d0ea302843391ded09fbb29d3ffece2abf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3907438
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>