cryptohome: Connect AuthenticateAuthFactor to the auth block util verify
This hooks up the new VerifyWithAuthFactorAsync API in the auth block
utility to the Authenticate code. In theory this means that you can
authenticate with legacy fingerprint, although we still need APIs to
enable and disable the fingerprint sensor in order for that to actually
work.

As you can see it somewhat duplicates what the credential verifier is
doing re: verification, but also see the TODO and bug for addressing
that in a followup.

BUG=b:246824625
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I2781285418d75f2d59c5e182f4413a13a769d796
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3911217
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>