installer: Convert BiosType to enum class
This is in preparation for using BiosType in UMAs. Enum UMAs are cleaner
and safer to use when you can supply a kMaxValue that matches the
highest value, but 'normal' `enum`s don't allow duplicate values. An
`enum class` is needed to support the duplicate value.

BUG=b:235257917
TEST=`FEATURES=test emerge-$BOARD chromeos-installer`

Change-Id: I0c6a0a08f1836aeeb9acc295a523b6ac10dd9fac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3910074
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>