installer: Add UMAs to postinstall
The reven board expands the variety of bios_types we expect to see in
use. The way to track that usage in ChromeOS is through UMAs.

Add:
* a USE flag to control whether to use metrics
* conditional inclusion of the metrics library
* a wrapper to support compilation and testing with/without libmetrics
* some UMA metrics

We only include the metrics library when the postinst_metrics flag is
specified to reduce the effect on boards that don't send metrics
during postinstall (which we expect to be most of them.)

The specific metrics:
* Installer.Postinstall.BiosType
For monitoring distribution of EFI/Legacy or other boot types

* Installer.Postinstall.NonChromebookBiosSuccess.*
For monitoring failure rates on bios types that take "special handling":
UBoot, EFI, Legacy

BUG=b:235257917
TEST=USE=-postinst_metrics FEATURES=test emerge-${B} chromeos-installer
TEST=USE=postinst_metrics FEATURES=test emerge-${B} chromeos-installer

Cq-Depend: chromium:3914606
Change-Id: Ied09190a2294563a976a00a64d347f3413c2f9ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3910075
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>