shill: Stop using 3-arity version of SLOG()
Compared with the 2-arity version, the 3-arity version can specify the
scope of this log statement, which can be a different scoped defined in
the file.

This patch removes the usage of the 3-arity version of SLOG() and thus
the affected SLOG() statements will be forced to follow the scope in the
current file.

Also introduces the 1-arity version of SLOG(). All SLOG() statements
without logging an dynamic ObjectID will be changed to use this one by
the follow-up patches.

BUG=b:248154452
TEST=FEATURES="test" emerge-${BOARD} shill shill-net

Disallow-Recycled-Builds: test-failures
Change-Id: I99b02ab4f6b799e0f2366921d5e03656e559d8bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3907306
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>