shill: Replace SLOG(nullptr, v) with SLOG(v)
Currently, the touched SLOG() statements in this patch  will output a
const string defined in ObjectID(), which does not provide any more
information given that we already log the filename. This patch changes
them to use the 1-arity version of SLOG() to avoid logging ObjectID()
directly.

This patch is generated by replacing "SLOG(this, " with "SLOG(" in the
files that the pointer is not used, and then fixing compile, style, and
unit_test issues manually.

This series of patches can reduce the binary size of shill by about 50KB
on octopus-kernelnext.

BUG=b:248154452
TEST=FEATURES="test" emerge-${BOARD} shill shill-net

Disallow-Recycled-Builds: test-failures
Change-Id: If553fec526eefcc9b9d39fa3270c1db8af381dd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3907307
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>