cryptohome: Support a const iterator for UserSessionMap
This basically requires rolling our own iterator wrapper, instead of
leveraging the iterator of the underlying std::map. Mainly because the
CrOS const correctness rules are stricter and incompatible with the
stdlib ones.

Note that in addition to the benefit of allowing use to iterator over
const UserSessionMap objects, this also means you now can't use the
iterator to modify the underlying map (e.g. by replacing or deleting the
UserSession object), even when you have a non-const iterator.

While doing this I also converted most of the for loops iterating over
this to use structured bindings.

BUG=b:243846478
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: If96cc8578fe45e455730c1d98e3805ae1be85493
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3914135
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>