shill: cellular: Default attach IP type to IPv4
If an attach APN does not have an IP type defined, shill does not send
the ip-type to ModemManager, and ModemManager does not allow this for
qmi modems. For mbim modems, MM allows no ip-type, but this leaves it
to the modem to choose an IP type. All modb APNs already default to
ipv4, but UI APNs don't do this.

BUG=b:248160274
TEST=Verify that the attach APN can be updated with a UI APN.

Change-Id: I22eeabad7793c0600cb28ae83df0e4305d3f6360
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3914132
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>