diagnostics: Change FetchCpuInfo to pure function
Change the structure of cpu fetcher such that a wrapper function is
called, and a cpu fetcher class is hidden from the caller.

Edit the cpu fetcher class such that a static void function |Fetch| is
called and an instance is created within the |Fetch| function. The
ownership of the object is then moved into a callback barrier.

BUG=b:248224656
TEST=(cr) FEATURES=test USE="-cros-debug" emerge-zork diagnostics
  (cr) FEATURES=test USE="-cros-debug" emerge-volteer diagnostics
  (DUT) cros-health-tool telem --category=cpu

Change-Id: Ifaa77b75f3f810fccd4bd4107b01a086c5ee6edc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3915655
Auto-Submit: Ethan Cheng <yycheng@google.com>
Commit-Queue: Ethan Cheng <yycheng@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Tested-by: Ethan Cheng <yycheng@google.com>
Reviewed-by: Chung-Sheng Wu <chungsheng@google.com>