fusebox: have client call ListStorages
Prior to this commit, there was potentially a small window of time, if
the FuseBoxService server (in the Chromium process) came up before the
FuseBoxService client (in the FuseBox process) where StorageAttached
signals could be broadcast with no client listening to hear them.

Calling ListStorages addresses that, so that the client is initialized
with the full list of attached storages even if it arrives second.

Also remove the obsolete --storage arg, which was used for bootstrapping
in early development (when the client held the subdir to FSURL prefix
map), but is currently ineffective (as the server now holds it).

BUG=chromium:1249754
TEST=manual

Change-Id: I4f5cdf4f4adec7243cf763d5be8bce06cab027d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3915659
Auto-Submit: Nigel Tao <nigeltao@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Nigel Tao <nigeltao@chromium.org>