shill: net: Replace SLOG() with VLOG()
Since shill-net is only compiled as a dynamic lib, SLOG() statements in
it all always compiled to VLOG(). This patch is just a cleanup and does
not have any impact on the final binary file (can be observed by no size
change for both shill and shill-net.so).

BUG=b:248154452
TEST=emerge-${BOARD} shill shill-net

Change-Id: Id804fe128f7161e2f5fb96624f1be524b82ec0c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3915720
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>