hwsec-utils: add exit-code-initializable Output type for unit tests
The original design implements expected exit status with
std::process::ExitStatus, which is incorrect - the parameter for from_raw
is not the exit status instead of the wait status. So, this CL add a new
data type HwsecOutput and replace the original Output with it.

BUG=None
TEST=build ok
TEST=hwsec-utils unittests

Change-Id: I4996a4a48a6a6f68a43fbd25f7623fccd2a69f55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3916069
Tested-by: Chris Chien <cychien@google.com>
Commit-Queue: Chris Chien <cychien@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Howard Yang <hcyang@google.com>