shill: cellular: Support filtering by GID1 on MODB
Add a new message filter type to allow filtering MVNOs and APNs by the
SIM EF_GID1 file contents.

The GID1 field is given by ModemManager as an array of bytes, e.g.:
{ 0xAF, 0x1D }, which will be converted by shill into an hex-encoded
ASCII string.

The GID1 field filter values are expected in the MODB associated to
either the "regex" or "exclude_regex" match types, specified as
regular expressions to match or not match a given hex-encoded ASCII
string. E.g. regex = "AF1D" can be used for an exact match, which is
the most usual case.

The "range" match type is not allowed with this filter type.

BUG=b:207400969
TEST=Common

Change-Id: Id27527db546e4b9cb1422843dc3fcaa83147c54e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3923334
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>