libec: Avoid overwriting request/response
The "Data" struct had a single buffer formed with a union of the request
and response. After running a command, this would result in the request
being overwritten. This commit makes a change to store the request and
response separately from the command buffer so that running the command
does not affect the request and modifying the request does not modify
the response.

BUG=b:248622515
TEST=FEATURES="test" emerge-hatch libec

Change-Id: I3a0f074950dce3fc88ef3a19b82aa1fdc0de63e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3919775
Reviewed-by: Andrea Grandi <agrandi@google.com>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>