cryptohome: Add version data to the SerializedAuthFactor flatbuffer
This adds new fields in order to capture the new ChromeOS version and
Chrome version information that was added to the AuthFactor common
metadata. The field definitions map onto the ones in the proto in a
pretty straightforward way.

This also changes the proto fields from bytes to string. This is,
strictly speaking, not a safe change if these fields were ever being
populated with bytes which are not valid UTF-8. However, since the
values of these versions are always supposed to be valid UTF-8, _and_
we're not actually using these fields yet, this should be acceptable.

BUG=b:240597941
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I7cdd4252f397819d39d72f60e65c60c93281ece0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3919772
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>