shill: Include service name in connection drop/failed to connect logs
This patch adds service name in the connection drop and failed to
connect logs. This will be helpful to easily determine the service
to which connection drop belongs to and should be straight forward
to parse from feedback report analyzer or any other script.

BUG=None
TEST=Build succeeds

Change-Id: Ice0b52ee7eab4f8c365001b678ad6b10ebf21d93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3922653
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>