cryptohome: Populate the AuthFactor common metadata
For the Chrome version, we simply add support for passing through the
version fields from proto to AuthFactor (and back). Built on top of the
existing support for serializing AuthFactors, it should be preserved.

For the ChromeOS version we do the same, but with the added hook that we
populate the version directly in cryptohome, rather than relying on any
passed-in values.

BUG=b:240597941
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I228ddc164ff5ebaef2ac1bdebeff30d38f465bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3924682
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>