patchpanel: ndproxy: turn on RA-injection for Ethernet and Wifi
This patch turns on the same RA injection we did on cellular for
Ethernet and Wifi as well, making guests treat ChromeOS host as router
and neighbors in local network as one hop away. This reduces the amount
of NS and NA to be sent to the physical network. (Note that neighbors
in the physical network still treat guests as on-link neighbors).

BUG=b:246444885
TEST=`tast run local arc.IPv6Connectivity`

Change-Id: Ieadc95cb2602fcc0bc60e05674c419ba8759d052
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3925235
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>