power: Fix suspend invocation logic error
https://crrev.com/c/3907727 fixed some linter errors, but regressed
suspend. It's counterintuitive, but for if statements, the contents
underneath the if are evaluated if the return code for its expression
is 0. So for instance, in the snippet below, true returns 0, and the
snippet prints "yay":

if true; then
  echo "yay"
else
  echo "boo"
fi

So the logic for evaluating invoke_hiberman_helper_process is backwards.
Let's fix it so all suspends don't report failure.

BUG=b:217670935
TEST=Attempt hibermation

Change-Id: Idac80afb610cac97bcee8d32479df104ee35f57a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3925728
Commit-Queue: Evan Green <evgreen@chromium.org>
Auto-Submit: Evan Green <evgreen@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sherri Lin <sherrilin@google.com>