bootstat: summary: Fix large number sorting
When the system has been up a while, the summary results would start
getting sorted very strangely. Example:

localhost ~ # bootstat_summary
    time %cpu       dt  %dt  event
89255900 100% 89255900 100%  network-ethernet-idle
89264000 100%     8100 100%  network-ethernet-registered
89266800 100%     2800 100%  network-ethernet-configuration
89267500 100%      700 100%  network-ethernet-ready
89267600 100%      100 100%  network-ethernet-online
    1228 100% -89266371 100%  pre-startup
    9445 100%     8216 100%  post-startup
   11466 100%     2021 100%  shill-start
...

This is because the first 'awk' wasn't specifying an output format for
printing numbers, and so we end up in scientific notation, like:

  $ echo 89255900.8761235 | awk '{print 1000*$1}'
  8.92559e+10

Numeric sort only understood the first part (8.92559) and believed that
to be smaller than 1228.

With an OFTM (output format) set, we ensure this is kept in decimal
format, not scientific notation, and the sorting works.

BUG=none
TEST=bootstat_summary on long-running system

Change-Id: Ia661688ccd14dafe2df60a830fe237aa062d307c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3928222
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Chun-Tse Shao <ctshao@google.com>