camera: remove fatal log in EglContext::GetSurfacelessContext
The LOGF(FATAL) generates crash reports when the system is in a really
bad state. The crash reports don't give us useful information for
debugging, but create a lot of noise. Besides, on (most) platforms
which don't require GPU acceleration, we don't have to crash when we
can't create the GL context.

BUG=b:247029133
TEST=Build and deploy cros-camera and cros-camera-libs. Verify that
camera works on brya and zork.

Change-Id: I3521d11cdb48b9634d55b0f17c6d9c7e67ef1521
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3925599
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>