cryptohome: Refactor verifier construction over to use AuthInput
In most of the codepaths this is a trivial refactor, since they were
just extracting the passkey from the AuthInput anyway. However, in the
legacy paths this changes things to hardcode in the use of a password
factor.

In practice this doesn't really change behavior, as there was already an
"unspecified -> password" assumption backed into the factory. However,
this instead pulls that logic up into the legacy paths, which has some
downsides (it now encodes the assumption in a couple places) but overall
has the benefit of leaving the credential verifier factory itself more
generic while confining the ugliness to the legacy code.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I258418aa340e536bfff57cb77840d5972acc56d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3926347
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>