cryptohome: AuthFactor auth for keydataless pass
Fix AuthenticateAuthFactor to work with legacy VaultKeysets
(those that don't have KeyData::label stored on disk) for password.

The bug was that the code was accidentally trying to load the VK using
the label from the parsed KeyData::label (which is empty in this case)
instead of the actual label (like "legacy-0").

BUG=b:249572472
TEST=existing unit tests

Change-Id: I740c0e0a7eedddc09ce7d4b2666c7c6c74c81ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3925631
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>