libec: Keep a copy of the command instead of Data
Now that we keep a copy of the request and response, it makes more sense
to only keep a copy of the command struct (cros_ec_command_v2) rather
than the entire Data struct. Similar to what is done with the request
and response, the command struct is copied when running a command.

BUG=b:248622515
TEST=FEATURES="test" emerge-hatch libec

Change-Id: I591c076685434a81f1e4caf7bd547ef66ce1c226
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3929466
Reviewed-by: Andrea Grandi <agrandi@google.com>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>