crash: KernelCollector: Check /sys/class/watchdog/watchdog0/bootstatus
The AMD sp5100 watchdog device correctly reports the boot status in the
kernel.

Update KernelCollector::LastRebootWasWatchdog() to check
/sys/class/watchdog/watchdog0/bootstatus for the last boot status. If a
watchdog reset occurred, collect a crash report with a signature that
contains the decoded bootstatus value.

If the file /sys/class/watchdog/watchdog0/bootstatus is not accessible
or reading the file fails, continue checking for the BIOS eventlog for
a watchdog reset.

This requires a call to minijail_bind() to make /sys/class/watchdog
accessible.

BUG=b:223617872
TEST=tast run <IP> crash.WatchdogCrash
TEST=cros_workon_make --test --board=guybrush crash-reporter

Change-Id: Iaf6eb8e9286c06729282ae9bbef62c2b6e287713
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3930818
Tested-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Tim Van Patten <timvp@google.com>