cryptohome: Get valid backup vault keyset after authentication
After completing the successful authentication with UserSecretStash
call GetValidKeyset() function on the same label to load the
decrypted backup VaultKeyset onto AuthSession's member |vault_keyset_|.
This will make the addition of a second backup VaultKeyset possible.

This is not a support for adding a PIN yet. Follow up CLs will modify
the reset secret generation with USS so that backup VKs will be able
have the same reset secret as the USS. And that will complete adding
a second backup VaultKeyset as PIN.

BUG=b:246656102
TEST=tast run IP cryptohome.* hwsec.Pin*

Change-Id: Ic2417e4670b7c9e805ccbe48b8b69aec3a11a2b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3933795
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>