camera: Update minijail to expose camera effect resources
The camera effects rely on resources stored on the disk to be able to
run. Currently the minijail in which the camera stack runs does not
mount the resources for the effects to work.

Add appropriate directory mount to minijail to make camera effect
resources available in the minijail

BUG=b:249396397
TEST=Manually ran tests on DUT (Brya) to demonstrate necessity of
including the /run/imageloader directory in the minijail

Change-Id: I1728868d703805b486ee8b577fef98ccc5fab6ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3933909
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Tom Shafron <shafron@google.com>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Reviewed-by: Tom Shafron <shafron@google.com>
Tested-by: Tom Shafron <shafron@google.com>