cryptohome: Allow sessions to have multiple verifiers.
This converts both auth and user session into storing the verifiers in a
std::map of label -> verifier, instead of just a solo unique_ptr.

In practice this doesn't change behavior all that much because we still
only do password-typed verifiers, but there are still some significant
changes: if you add multiple password auth factors you get multiple
verifiers!

The interaction with the legacy keydata paths gets a bit weird, because
we still only store a single KeyData object in the user session and so
the old paths that still use that directly just get "one verifier" in
that sense. I try to preserve the same behavior with that as well: if
you add a second factor it doesn't overwrite the KeyData from the first
one, and if you remove the factor whose label matches the KeyData label
then we also clear the KeyData. We still be able to rip all that code
out when we drop those legacy paths, though.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I721c0e8724c0d72f702edfc72c1a319001efa993
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3935362
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>