shill: Configure wpa supplicant sae_check_mfp option
There are some flex devices that do not support PMF but
do support SAE due to missing group management ciphers.
APs in mixed mode will reject association when SAE is
enabled but PMF is not. Enable the `sae_check_mfp` option
for networks to allow wpa_supplicant to limit SAE when
the device does not have the necessary ciphers.

With this extra check enabled devices that are capable
of WPA3 will fall back to PSK if the AP does not advertise
PMF for the connection. This makes it comply with the
WPA3 specification v3.0 section 2.3 item 8 "A STA shall negotiate PMF when associating to an AP using SAE".

BUG=b:238933377
TEST=connect to wpa2+wpa3 mixed mode AP on the device without BIP

Cq-Depend: chromium:3901830
Change-Id: Ia67bbfa568ec8bbbc7d06f137a9f3cda1b579cdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3932696
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Tested-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>