shill: Stop using std::optional for IPConfig::Properties in Network
std::optional is an object but not a pointer so it is heavy in space
when the original class is large. This patch changes the
link_protocol_ipv4_properties and ipv6_static_properties in Network from
std::optional to std::unique_ptr since they are not used by every
Network object and can also avoid the additional copy in the setter
functions.

BUG=b:232177767
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=tast run $DUT network.VPNConnect.*

Change-Id: I722968b3919371e3d3b1917faf6d0d58ad98924e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3934077
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>