shill: cellular: Add last connected default/attach APN properties
Add LastConnectedDefaultAPN and LastConnectedAttachAPN to
CellularService, which track the last default and attach APNs that were
used for connection. These properties are similar to LastGoodAPN and
LastAttachAPN, except they are not cleared when the connection fails.
These properties will be used by Chrome for migration  when the
ApnRevamp flag is first enabled, in order to preserve custom APNs used
when the flag was disabled.

BUG=b:162365553
TEST=CellularServiceTest

Change-Id: I773b55704d0ea47070f5cab37ac2989761149abf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3936683
Commit-Queue: Gordon Seto <gordonseto@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Gordon Seto <gordonseto@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>