init: Make sure STATE_DEV and DEV_IMAGE are set on resume
For hibernate resume boots we want to make sure that STATE_DEV
and DEV_IMAGE are properly set for use later in the script. The
C++ code is already calling hiberman resume-init so calling it
again is likely not going to work as intended. Assume that
the /dev/mapper/ paths will only be created in the successful
resume-init case, so use that to determine if we should set
STATE_DEV and DEV_IMAGE accordingly.

BUG=b:172210918
TEST=build_image --board=brya-hibernate

Change-Id: I272c883b125d53cfc3d0ff206dc604a79f8c5b6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3936675
Commit-Queue: Alexis Savery <asavery@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>