libchromeos-rs: migrate sys_util::syslog
os_install_service uses features of sys_util::syslog that are not
present in the syslog crate so migrate sys_util::syslog to
libchromeos::deprecated.

The unit tests for sys_util::syslog were broken because they depend on
the actual syslog, so new_seqpacket_pair was brought over from crosvm as
a way of making a mock syslog connection for the unit tests.

BUG=b:250883877
TEST=cargo test

Change-Id: I5e6e390735d166d1dfb1d3b796c691da8d514b39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3937624
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Paramjit Oberoi <psoberoi@google.com>