cryptohome: Push verifiers from AuthSession directly to UserSession
This significantly reworks the flow of CredentialVerifier instances from
AuthSession -> UserSession. Instead of storing verifiers in AuthSession
and then relying on UserSession grabbing (or being asked to grab) them
from the AuthSession at any necessary points, we instead have
AuthSession send all verifiers directly to the UserSessionMap via a
Forwarder object it creates.

The trickiness is necessary because a UserSession doesn't necessarily
exist at the point where the AuthSession does, so it can't just send the
verifiers directly to the UserSession. So instead we use this forwarding
object that can hold on to the verifiers and automatically pass them on
when the UserSession gets added.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I73275a771993b954c3dc593e89fc009a7ee1359e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3936027
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>