faced: Add QueueingStream data type
The `QueueingStream` data type is a queue that decouples writers from
readers, and allows readers to asynchronously wait for new items to
be added to the queue.

Our intended use is to allow frames to be written by a camera object
and read as required by the enrollment and authentication classes.

BUG=b:237233809
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/faced`

Change-Id: I889bf80b24a4efe26bc67cdf8686cab9e7159826
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3941326
Reviewed-by: Victor Wong <vwon@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Ben Kersten <bkersten@google.com>
Tested-by: David Greenaway <dgreenaway@google.com>